requirements.txt
python-dotenv
numpy
pandas
psycopg2
aiohttp
tiktoken
fastapi
pydantic
uvicorn